chore: combine two scripts into one
